app/requirements.txt
mem0ai>=0.1.50
fastmcp>=2.0.0
openai>=1.50.0
qdrant-client>=1.9.0
